optimization: 优化Tag-input组件的代码结构
